extract part of wizard criteria page into ExplorerComponent
