+++ ui/src/app/cohort-search/criteria-wizard/wizard/wizard.component.css
@@ -29,7 +29,7 @@
   flex-flow: column;
 }
 
-#criteria-page > .criteria-tree-container {
+#criteria-explorer {
   width: 100%;
   height: 70%;
   overflow-y: auto;
@@ -38,7 +38,7 @@
   position: relative;
 }
 
-#criteria-page > .criteria-selection-container {
+#criteria-selection {
   width: 100%;
   height: 30%;
   flex: 3;
@@ -51,11 +51,11 @@
   #criteria-page {
     flex-flow: row;
   }
-  #criteria-page > .criteria-tree-container {
+  #criteria-explorer{
     width: 70%;
     height: 100%;
   }
-  #criteria-page > .criteria-selection-container {
+  #criteria-selection {
     width: 30%;
     height: 100%;
   }
